+++ provider/pkg/resources/custom_storage_test.go
@@ -1,0 +1,19 @@
+// Copyright 2016-2023, Pulumi Corporation.
+
+package resources
+
+import (
+	"github.com/stretchr/testify/assert"
+	"testing"
+)
+
+func TestParseBlobIdProperties(t *testing.T) {
+	id := "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/myrg/providers/Microsoft.Storage/storageAccounts/mysa/blobServices/default/containers/myc/blobs/folder%2Flog.txt"
+	props, ok := parseBlobIdProperties(id)
+	assert.Equal(t, ok, true)
+	assert.Equal(t, len(props), 4)
+	assert.Equal(t, props["resourceGroupName"].StringValue(), "myrg")
+	assert.Equal(t, props["accountName"].StringValue(), "mysa")
+	assert.Equal(t, props["containerName"].StringValue(), "myc")
+	assert.Equal(t, props["blobName"].StringValue(), "folder%2Flog.txt")
+}